+++ hpobench/benchmarks/surrogates/paramnet_benchmark.py
@@ -44,6 +44,16 @@ cd /path/to/HPOBench
 pip install .[paramnet]
 
 ```
+
+Changelog:
+==========
+0.0.2:
+* Fix OnTime Test function:
+  The `objective_test_function` of the OnTime Benchmarks now checks if the budget is the right maximum budget.
+* Standardize the structure of the meta information
+
+0.0.1:
+* First implementation
 """
 
 import logging
@@ -55,7 +65,7 @@ import numpy as np
 from hpobench.abstract_benchmark import AbstractBenchmark
 from hpobench.util.data_manager import ParamNetDataManager
 
-__version__ = '0.0.1'
+__version__ = '0.0.2'
 
 logger = logging.getLogger('Paramnet')
 
@@ -276,10 +286,13 @@ class _ParamnetOnTimeBenchmark(_ParamnetBase):
     def objective_function_test(self, configuration: Union[CS.Configuration, Dict],
                                 fidelity: Union[CS.Configuration, Dict, None] = None,
                                 rng: Union[np.random.RandomState, int, None] = None, **kwargs) -> Dict:
+        # Get the maximum fidelity for this benchmark.
+        fidelity_space = self.get_fidelity_space()
+        max_budget = fidelity_space.get_hyperparameter('budget').upper
 
-        assert fidelity['step'] == 50, f'Only querying a result for the 50. epoch is allowed, ' \
-                                       f'but was {fidelity["step"]}.'
-        return self.objective_function(configuration, fidelity={'step': 50}, rng=rng)
+        assert fidelity['budget'] == max_budget, f'Only querying a result for a budget of {max_budget} is allowed, ' \
+                                                 f'but was {fidelity["budget"]}.'
+        return self.objective_function(configuration, fidelity={'budget': max_budget}, rng=rng)
 
     @staticmethod
     def _get_fidelity_space(dataset: str, seed: Union[int, None] = None) -> CS.ConfigurationSpace: